some fixes to android build
git-svn-id: file:///Users/steve/svn/ckfft/trunk@77 743b922c-e3f2-40ca-8674-e6e14afc19ac

+++ src/test/ios/main.mm
@@ -14,7 +14,7 @@
 void* testThreadProc(void*)
 {
     int val = test() ? 0 : 1;
-    return (void*) val;
+    return (void*) (size_t) val;
 }
 
 #define PTHREAD_VERIFY(x) { int ret = x; assert(ret == 0); }
@@ -36,7 +36,7 @@ int startTest()
 
     void* ret;
     pthread_join(threadId, &ret);
-    return (int) ret;
+    return (int) (size_t) ret;
 }
 
 int main(int argc, char *argv[])
